Move util methods and structs from swap to chain
